chore: Rename Minikube registry discovery package
